refactored packit/git_utils.py into packit/utils/repo.py
